Очищены шаблоны, удалены не нужные шаблоны
