--- apps/gallery/templates/album/create.html
@@ -1,19 +1,0 @@
-{% extends 'base.html' %}
-{% load crispy_forms_tags %}
-
-{% block title %}Галерея | Создание альбома{% endblock %}
-
-{% block content %}
-    <div class="row">
-        <div class="col-sm-12">
-            <h3 class="card-title text-center my-5">Добавить альбом</h3>
-            <form class="w-75 mx-auto" action="" method="post" enctype="multipart/form-data">
-                {% csrf_token %}
-                {{ form.media }}
-                {{ form|crispy }}
-                <a href="{% url 'gallery:album_list' %}" class="btn btn-secondary">Отмена</a>
-                <button type="submit" class="btn btn-primary">Сохранить</button>
-            </form>
-        </div>
-    </div>
-{% endblock %}
