+++ apps/gallery/templates/album/detail.html
@@ -5,41 +5,10 @@
 
 {% block content %}
     <div class="row justify-content-center">
-        <div class="col-1">
-            <a href="{% url 'gallery:album_list' %}" class="btn btn-secondary" data-bs-toggle="tooltip"
-               data-bs-placement="top"
-               data-bs-title="Вернуться к альбомам">
-                <i class="fa-solid fa-arrow-rotate-left"></i>
-            </a>
-        </div>
         <div class="col-3">
             <h3 class="text-center">{{ album.title }}</h3>
         </div>
     </div>
-    {% if request.user.is_staff %}
-        <div class="row">
-            <div class="col-12 mx-auto mt-5 text-center">
-                <span class="text-muted mr-3">ADMIN:</span>
-                <a href="{{ album.get_deletion_to_cart_url }}" class="btn btn-warning" data-bs-toggle="tooltip"
-                   data-bs-placement="top" data-bs-title="Удалить в корзину">
-                    <i class="fa-solid fa-trash"></i>
-                </a>
-
-                <a href="{{ album.get_deletion_url }}" class="btn btn-danger" data-bs-toggle="tooltip"
-                   data-bs-placement="top" data-bs-title="Удалить окончательно">
-                    <i class="fa-solid fa-minus"></i>
-                </a>
-                <a href="{{ album.get_update_url }}" class="btn btn-info" data-bs-toggle="tooltip"
-                   data-bs-placement="top" data-bs-title="Редактировать">
-                    <i class="fa-solid fa-pen"></i>
-                </a>
-                <a href="{% url 'gallery:content_create' album.id %}" class="btn btn-primary" data-bs-toggle="tooltip"
-                   data-bs-placement="top" data-bs-title="Добавить изображения">
-                    <i class="fa-solid fa-plus"></i>
-                </a>
-            </div>
-        </div>
-    {% endif %}
     <div class="row">
         {% if not videos and not images %}
             <p class="text-center mt-5">Контента еще нет :)</p>
