+++ apps/gallery/templates/album/list.html
@@ -4,16 +4,6 @@
 
 {% block content %}
     <h3 class="text-center">Галерея</h3>
-    {% if request.user.is_staff %}
-        <div class="row">
-            <div class="col-12 mx-auto mb-5 text-center">
-                <span class="text-muted mr-3">ADMIN:</span>
-                <a href="{% url 'gallery:album_create' %}" class="btn btn-primary" data-bs-toggle="tooltip" data-bs-placement="top" data-bs-title="Добавить альбом">
-                    <i class="fa-solid fa-plus"></i>
-                </a>
-            </div>
-        </div>
-    {% endif %}
     {% if albums %}
         <div class="row">
             <div class="grid mx-auto" data-masonry='{ "itemSelector": ".grid-item", "isFitWidth": true }'>
