--- apps/gallery/templates/album/update.html
@@ -1,19 +1,0 @@
-{% extends 'base.html' %}
-{% load crispy_forms_tags %}
-
-{% block title %}
-    Галерея | Правка альбома
-{% endblock %}
-
-{% block content %}
-    <div class="row">
-        <h3 class="text-center mb-5">Редактирование альбома</h3>
-        <form method="post">
-            {% csrf_token %}
-            {{ form|crispy }}
-            <a href="{% url 'gallery:album_detail' request.resolver_match.kwargs.pk %}" class="btn btn-secondary">Отмена</a>
-            <button type="submit" class="btn btn-success">Сохранить</button>
-        </form>
-    </div>
-
-{% endblock %}
