--- apps/gallery/templates/content/create.html
@@ -1,17 +1,0 @@
-{% extends 'base.html' %}
-{% load crispy_forms_tags %}
-
-{% block content %}
-    <div class="row">
-        <div class="col-sm-12">
-            <h3 class="card-title text-center my-5">Добавить изображения</h3>
-            <form class="w-75 mx-auto" action="" method="post" enctype="multipart/form-data">
-                {% csrf_token %}
-                {{ form.media }}
-                {{ form|crispy }}
-                <a href="{% url 'gallery:album_detail' request.resolver_match.kwargs.album_pk %}" class="btn btn-secondary">Отмена</a>
-                <button type="submit" class="btn btn-primary">Сохранить</button>
-            </form>
-        </div>
-    </div>
-{% endblock %}
