--- apps/gallery/templates/content/delete.html
@@ -1,26 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}
-    Публикации | Удаление
-{% endblock %}
-
-{% block content %}
-    <div class="row">
-        <div class="col-sm-12 col-md-6">
-            <img src="{{ albumcontent.file.url }}" width="400" alt="">
-        </div>
-        <div class="col-sm-12 col-md-6">
-            <div class="card mx-auto">
-                <div class="card-body text-center">
-                    <p class="card-text">Удалить контент?</p>
-                    <form action="" method="post" enctype="multipart/form-data">
-                        {% csrf_token %}
-                        <a href="{{ albumcontent.get_absolute_url }}" class="btn btn-secondary">Отмена</a>
-                        <button type="submit" class="btn btn-danger">Удалить</button>
-                    </form>
-                </div>
-            </div>
-        </div>
-    </div>
-
-{% endblock %}
