--- apps/gallery/templates/content/detail.html
@@ -1,59 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}Галерея | Альбом | Информация об изображении{% endblock %}
-
-{% block content %}
-    <div class="row justify-content-center">
-        <div class="col-1">
-            <a href="{% url 'gallery:album_detail' request.resolver_match.kwargs.album_pk %}" class="btn btn-secondary" data-bs-toggle="tooltip"
-               data-bs-placement="top"
-               data-bs-title="Вернуться к альбомам">
-                <i class="fa-solid fa-arrow-rotate-left"></i>
-            </a>
-        </div>
-        <div class="col-8">
-            <h3 class="text-center">Информация об изображении</h3>
-        </div>
-    </div>
-    {% if request.user.is_staff %}
-        <div class="row">
-            <div class="col-12 mx-auto mt-5 text-center">
-                <span class="text-muted mr-3">ADMIN:</span>
-                <a href="{{ albumcontent.get_deletion_to_cart_url }}" class="btn btn-warning" data-bs-toggle="tooltip"
-                   data-bs-placement="top" data-bs-title="Удалить в корзину">
-                    <i class="fa-solid fa-trash"></i>
-                </a>
-                <a href="{{ albumcontent.get_deletion_url }}" class="btn btn-danger" data-bs-toggle="tooltip"
-                   data-bs-placement="top" data-bs-title="Удалить окончательно">
-                    <i class="fa-solid fa-minus"></i>
-                </a>
-                <a href="{{ albumcontent.get_update_url }}" class="btn btn-info" data-bs-toggle="tooltip"
-                   data-bs-placement="top" data-bs-title="Редактировать">
-                    <i class="fa-solid fa-pen"></i>
-                </a>
-            </div>
-        </div>
-    {% endif %}
-    <div class="row mt-5 mx-auto">
-        <div class="col-sm-12 col-md-6">
-            <div class="card">
-                <img src="{{ albumcontent.file.url }}" width="400" alt="" class="card-img-top">
-            </div>
-        </div>
-        <div class="col-sm-12 col-md-6 align-items-end">
-            <ul class="list-group">
-                <li class="list-group-item">Загружено: {{ albumcontent.created_at }}</li>
-                <li class="list-group-item">Последнее обновление: {{ albumcontent.updated_at }}</li>
-                <li class="list-group-item">Ширина: {{ albumcontent.file.width }}</li>
-                <li class="list-group-item">Высота: {{ albumcontent.file.width }}</li>
-                <li class="list-group-item">
-                    Видимость: {% if albumcontent.is_visible %}Да{% else %}Нет{% endif %}
-                </li>
-                <li class="list-group-item">
-                    Изображение удалено: {% if albumcontent.is_deleted %}Да{% else %}Нет{% endif %}
-                </li>
-            </ul>
-
-        </div>
-    </div>
-{% endblock %}
